ToDo/requirements.txt
Django==2.1.5
django-crispy-forms==1.7.2
gunicorn==19.6.0
html5lib==1.0.1
whitenoise==3.2.1
dj-database-url==0.5.0
psycopg2==2.7.7
pytz==2018.4
django-bootstrap-modal-forms==1.4.1
django-widget-tweaks==1.4.2
django-allauth==0.39.1

